Add files via upload
User can select the attributes he wishes to perform analysis on.

Acceptance Criteria
In order to use the analysis functionality the user has to fill the select boxes corresponding to the metrics needed to perform the chosen analysis, using the drop down menu which would contain the names of column names from the data set uploaded. All the required fields must be filled without any duplicate entries for the analysis to begin.