Add files via upload
This analysis answers the following question:
At the current selling price of the products, what is the expected demand for the upcoming year? (It would help the user to keep the stock of products accordingly in his inventory and also would help him achieve undisturbed sales.)

Acceptance Criteria
Using the selected fields against the required metrics for the analysis, the expected demand should be calculated and accordingly suggestions for the quantities of various products to be kept in future should be made.